requirements.txt
appdirs==1.4.3
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
coloredlogs==14.0
cssselect==1.1.0
distlib==0.3.0
docopt==0.6.2
eutils==0.6.0
filelock==3.0.12
habanero==0.7.2
humanfriendly==8.2
idna==2.9
importlib-metadata==1.6.0
lxml==4.5.0
metapub==0.5.5
more-itertools==8.2.0
mysql-connector-python==8.0.20
neo4j==1.7.6
neobolt==1.7.17
neotime==1.7.4
packaging==20.3
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
python-Levenshtein==0.12.0
pytz==2020.1
requests==2.23.0
six==1.14.0
tabulate==0.8.7
toml==0.10.0
tox==3.15.0
tqdm==4.46.0
Unidecode==1.1.1
urllib3==1.25.9
virtualenv==20.0.20
wcwidth==0.1.9
xlrd==1.2.0
zipp==3.1.0
